+++ frontend/src/app/pages/exercises/exercise/shared/exercise-map/feature-managers/moveable-feature-manager.ts
@@ -157,7 +157,8 @@ export abstract class MoveableFeatureManager<
 
     public onFeatureClicked(
         event: MapBrowserEvent<any>,
-        feature: Feature<FeatureType>
+        feature: Feature<FeatureType>,
+        condition?: boolean
         // eslint-disable-next-line @typescript-eslint/no-empty-function
     ): void {}