+++ frontend/src/app/pages/landing-page/landing-page/landing-page.component.ts
@@ -1,18 +1,28 @@
-import { Component, inject } from '@angular/core';
+import { Component, inject, resource, signal } from '@angular/core';
 import { Router, RouterLink } from '@angular/router';
-import type { ExportImportFile } from 'fuesim-digital-shared';
+import {
+    isAccessKey,
+    isExerciseKey,
+    isGroupParticipantKey,
+    type ExportImportFile,
+} from 'fuesim-digital-shared';
 import { escapeRegExp } from 'lodash-es';
 import { FormsModule } from '@angular/forms';
+import {
+    form,
+    FormField,
+    validate,
+    validateAsync,
+} from '@angular/forms/signals';
 import { AuthService } from '../../../core/auth.service';
 import { ApiService } from '../../../core/api.service';
 import { MessageService } from '../../../core/messages/message.service';
 import { HeaderComponent } from '../../../shared/components/header/header.component';
-import { JoinIdDirective } from '../../../shared/validation/join-id-validator.directive';
-import { ExerciseExistsValidatorDirective } from '../../../shared/validation/exercise-exists-validator.directive';
 import { AutofocusDirective } from '../../../shared/directives/autofocus.directive';
 import { DisplayValidationComponent } from '../../../shared/validation/display-validation/display-validation.component';
 import { FileInputDirective } from '../../../shared/directives/file-input.directive';
 import { FooterComponent } from '../../../shared/components/footer/footer.component';
+import { DisplayModelValidationComponent } from '../../../shared/validation/display-model-validation/display-model-validation.component';
 
 @Component({
     selector: 'app-landing-page',
@@ -21,13 +31,13 @@ import { FooterComponent } from '../../../shared/components/footer/footer.compon
     imports: [
         HeaderComponent,
         FormsModule,
-        JoinIdDirective,
-        ExerciseExistsValidatorDirective,
         AutofocusDirective,
         DisplayValidationComponent,
         RouterLink,
         FileInputDirective,
         FooterComponent,
+        FormField,
+        DisplayModelValidationComponent,
     ],
 })
 export class LandingPageComponent {
@@ -38,18 +48,62 @@ export class LandingPageComponent {
 
     public loginUrl = this.auth.loginUrl;
 
-    public exerciseId = '';
-
     public exerciseHasBeenCreated = false;
 
     public trainerKey = '';
-
     public participantKey = '';
+
+    readonly model = signal<{ joinKey: string }>({
+        joinKey: '',
+    });
+    readonly joinForm = form(this.model, async (schemaPath) => {
+        validate(schemaPath.joinKey, ({ value }) => {
+            if (!isAccessKey(value())) {
+                return {
+                    kind: 'no_access_key',
+                    message:
+                        'Die Eingabe muss eine gültige Übungs-PIN (6, 7 oder 8 Ziffern) sein.',
+                };
+            }
+            return null;
+        });
+        validateAsync(schemaPath.joinKey, {
+            params: ({ value }) => value(),
+            factory: (key) =>
+                resource({
+                    params: key,
+                    loader: async ({ params }) => {
+                        if (isExerciseKey(params)) {
+                            return (
+                                await this.apiService.exerciseExists(params)
+                            ).exists;
+                        } else if (isGroupParticipantKey(params)) {
+                            return (
+                                await this.apiService.parallelExerciseExists(
+                                    params
+                                )
+                            ).exists;
+                        }
+                        return false;
+                    },
+                }),
+            onSuccess: (result: boolean) => {
+                if (!result) {
+                    return {
+                        kind: 'username_taken',
+                        message: `Es existiert keine Übung mit dieser PIN.`,
+                    };
+                }
+                return null;
+            },
+            onError: () => null,
+        });
+    });
 
     public async createExercise() {
         this.apiService.createExercise().then((exerciseKeys) => {
             this.trainerKey = exerciseKeys.trainerKey;
-            this.exerciseId = this.trainerKey;
+            this.model.set({ joinKey: this.trainerKey });
             this.participantKey = exerciseKeys.participantKey;
             this.exerciseHasBeenCreated = true;
 
@@ -80,7 +134,7 @@ export class LandingPageComponent {
                     const exerciseKeys =
                         await this.apiService.importExercise(importPlain);
                     this.trainerKey = exerciseKeys.trainerKey;
-                    this.exerciseId = this.trainerKey;
+                    this.model.set({ joinKey: this.trainerKey });
                     this.participantKey = exerciseKeys.participantKey;
                     this.exerciseHasBeenCreated = true;
 
@@ -110,21 +164,29 @@ export class LandingPageComponent {
         }
     }
 
-    public pasteExerciseId(event: ClipboardEvent) {
+    public pasteExerciseKey(event: ClipboardEvent) {
         const pastedText = event.clipboardData?.getData('text') ?? '';
         const joinUrl = new RegExp(
-            `^${escapeRegExp(location.origin)}/exercises/(\\d{6,8})$`,
+            `^${escapeRegExp(location.origin)}/exercises.*/(\\d{6,8})$`,
             'u'
         );
 
         const matches = joinUrl.exec(pastedText);
         if (matches?.[1]) {
-            this.exerciseId = matches[1];
+            this.model.set({ joinKey: matches[1] });
             event.preventDefault();
         }
     }
 
     public joinExercise() {
-        this.router.navigate(['/exercises', this.exerciseId]);
+        const key = this.model().joinKey;
+        if (isExerciseKey(key)) {
+            this.router.navigate(['/exercises', this.model().joinKey]);
+        } else if (isGroupParticipantKey(key)) {
+            this.router.navigate([
+                '/exercises/parallel/join',
+                this.model().joinKey,
+            ]);
+        }
     }
 }