+++ frontend/src/app/shared/validation/custom-validators.ts
@@ -20,10 +20,4 @@ export namespace CustomValidators {
                 ? null
                 : { integer: true as const };
     }
-    export function joinIdValidator() {
-        return (control: AbstractControl) =>
-            /^((\d{6})|(\d{8}))$/u.test(control.value)
-                ? null
-                : { joinId: true as const };
-    }
 }