+++ shared/src/models/index.ts
@@ -26,5 +26,7 @@ export * from './operational-section.js';
 export * from './material-template.js';
 export * from './personnel-template.js';
 export * from './restricted-zone.js';
+export * from './scoutable.js';
+export * from './user-generated-content.js';
 export * from './technical-challenge/index.js';
 export * from './task.js';