+++ shared/src/state.ts
@@ -35,6 +35,10 @@ import {
     type ExerciseType,
     type ExerciseStatus,
     logEntrySchema,
+    scoutableSchema,
+    userGeneratedContentSchema,
+    Scoutable,
+    UserGeneratedContent,
     technicalChallengeSchema,
     TechnicalChallenge,
     taskSchema,
@@ -242,6 +246,19 @@ export class ExerciseState {
     @Equals(undefined)
     public previousTreatmentAssignment?: TreatmentAssignment;
 
+    @IsZodSchema(z.record(scoutableSchema.shape.id, scoutableSchema))
+    public readonly scoutables: { readonly [key: UUID]: Scoutable } = {};
+
+    @IsZodSchema(
+        z.record(
+            userGeneratedContentSchema.shape.id,
+            userGeneratedContentSchema
+        )
+    )
+    public readonly userGeneratedContents: {
+        readonly [key: UUID]: UserGeneratedContent;
+    } = {};
+
     /**
      * @deprecated Use {@link create} instead.
      */
@@ -256,5 +273,5 @@ export class ExerciseState {
      *
      * This number MUST be increased every time a change to any object (that is part of the state or the state itself) is made in a way that there may be states valid before that are no longer valid.
      */
-    static readonly currentStateVersion = 49;
+    static readonly currentStateVersion = 50;
 }