+++ shared/src/store/action-reducers/action-reducers.ts
@@ -20,6 +20,8 @@ import { RadiogramActionReducers } from './radiogram.js';
 import { VehicleTemplateActionReducers } from './vehicle-templates.js';
 import { RestrictedZoneActionReducers } from './restricted-zone.js';
 import { OperationalSectionActionReducers } from './operational-section.js';
+import { ScoutableActionReducers } from './scoutable.js';
+import { UserGeneratedContentActionReducers } from './user-generated-content.js';
 import { TechnicalChallengeActionReducers } from './technical-challenge.js';
 
 /**
@@ -47,6 +49,8 @@ const actionReducers = {
     ...VehicleTemplateActionReducers,
     ...RestrictedZoneActionReducers,
     ...OperationalSectionActionReducers,
+    ...ScoutableActionReducers,
+    ...UserGeneratedContentActionReducers,
     ...TechnicalChallengeActionReducers,
 } as const;