+++ shared/src/store/action-reducers/index.ts
@@ -10,6 +10,8 @@ export * from './action-reducers.js';
 export * from './vehicle-templates.js';
 export * from './operational-section.js';
 export * from './restricted-zone.js';
+export * from './scoutable.js';
+export * from './user-generated-content.js';
 export * from './technical-challenge.js';
 
 // Don't forget to register the actionReducer in `action-reducers.ts`